Fix import issue and move function
